전역 DB 예외처리 추가
