+++ src/features/companies/services/insertCompany.tsx
@@ -1,4 +1,5 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query'
+import { PostgrestError } from '@supabase/supabase-js'
 import supabase from '@/utils/supabase/client'
 import { useToast } from '@/hooks/use-toast'
 import { CompanySupabase } from '../data/schema'
@@ -13,6 +14,9 @@ export const insertCompany = async (newCompany: {
     .select('*')
     .single()
   if (error) {
+    if (error.code) {
+      throw new PostgrestError(error)
+    }
     throw error
   }
   return data
@@ -33,26 +37,6 @@ export const useInsertCompanyMutation = () => {
         description: (
           <pre className='mt-2 w-[340px] rounded-md bg-slate-950 p-4'>
             <code className='text-white'>{JSON.stringify(data, null, 2)}</code>
-          </pre>
-        ),
-      })
-    },
-    onError: (error) => {
-      console.error('Error inserting company:', error)
-
-      let errorMessage = '알 수 없는 오류가 발생했습니다.'
-      if (error.message) {
-        errorMessage = error.message
-      }
-
-      toast({
-        variant: 'destructive',
-        title: '데이터 삽입 실패!',
-        description: (
-          <pre className='mt-2 w-[340px] max-w-full overflow-x-auto rounded-md bg-slate-950 p-4'>
-            <code className='whitespace-pre-wrap text-white'>
-              {errorMessage}
-            </code>
           </pre>
         ),
       })
