+++ src/main.tsx
@@ -35,18 +35,7 @@ const queryClient = new QueryClient({
       staleTime: 10 * 1000, // 10s
     },
     mutations: {
-      onError: (error) => {
-        handleServerError(error)
-
-        if (error instanceof AxiosError) {
-          if (error.response?.status === 304) {
-            toast({
-              variant: 'destructive',
-              title: 'Content not modified!',
-            })
-          }
-        }
-      },
+      onError: handleServerError,
     },
   },
   queryCache: new QueryCache({
