+++ src/utils/postgrestErrorCodes.json
@@ -1,0 +1,28 @@
+{
+  "22003": "(숫자 범위 초과)",
+  "22004": "(NOT NULL 제약조건 위반)",
+  "22007": "(날짜/시간 형식 오류)",
+  "22012": "(0으로 나눔)",
+  "2201B": "(잘못된 정규 표현식)",
+  "22032": "(잘못된 json 문자열)",
+  "22039": "(잘못된 json 문자열)",
+  "2203A": "(잘못된 json 문자열)",
+  "2203B": "(잘못된 json 문자열)",
+  "2203C": "(잘못된 json 문자열)",
+  "2203D": "(잘못된 json 문자열)",
+  "2203E": "(잘못된 json 문자열)",
+  "2203F": "(잘못된 json 문자열)",
+  "2203G": "(잘못된 json 문자열)",
+  "23500": "(제약조건 위반)",
+  "23503": "(참조 무결성 제약조건 위반)",
+  "23505": "(고유 제약조건 위반)",
+  "23514": "(체크 제약조건 위반)",
+  "39004": "(NULL 값은 허용하지 않음)",
+  "3F000": "(스키마 없음)",
+  "42804": "(데이터 타입 불일치)",
+  "53100": "(디스크 공간 부족)",
+  "53200": "(시스템 메인 메모리 부족)",
+  "53300": "(너무 많은 연결)",
+  "58000": "(시스템 오류)",
+  "58030": "(I/O 오류)"
+}
